Move dependencies to their own directory.
