+++ lua/colorizer/trie.lua
